// ClientApp/src/Types/IProductDetails.ts
export default interface IProductDetails {
  activatedDate: string;
  branchID: string;
  branchesEmail: string;
  branchesName: string;
  branchesPhonenumber: string;
  code: string;
  descriptions: string;
  fromAddress: string;
  fromPlace: string;
  fromPlace_Name: string;
  isActive: boolean;
  name: string;
  productID: string;
  subcategoryID: number;
  toAddress: string;
  toPlace: string;
  toPlace_Name: string;
}